+++ lib/rails_multitenant/multitenant_model.rb
@@ -18,7 +18,9 @@ module RailsMultitenant
         scope_sym = "from_current_#{context_entity}".to_sym
 
         scope scope_sym, -> do
-          where(context_entity_id_field => GlobalContextRegistry[context_entity_id_field])
+          unless GlobalContextRegistry.use_unscoped_queries?
+            where(context_entity_id_field => GlobalContextRegistry[context_entity_id_field])
+          end
         end
 
         default_scope { send(scope_sym) }